requirements.txt
langchain
langchain-ollama
pydantic
ase
numpy
fairchem-core
huggingface_hub
